pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "servco-helpers"
version = "0.1.8"
description = "Internal helper utilities for tech projects"
authors = [{ name = "Albie Ewins" }]
readme = "README.md"
license = { text = "Proprietary - All Rights Reserved" }
requires-python = ">=3.9"

# Dependencies your helpers need
dependencies = [
    "google-cloud-storage",
    "streamlit",
    "streamlit-authenticator",
    "streamlit_autorefresh",
    "pandas",
    "numpy",
    "pyyaml",
    "servicepytan @ git+https://github.com/AE-servco/servicepytan.git@master",
    "google-cloud-storage",
    "google-cloud-secret-manager",
    "oauth2client",
    "gspread",
    "google-auth",
    "supabase",
    "uuid"
]

# Optional: nice metadata for documentation tools
keywords = ["helpers", "gcp", "utilities"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/AE-servco/servco-helpers"

[tool.setuptools]
packages = { find = { where = ["servco_helpers"] } }
